Adds internationalization to templates
Applies internationalization (i18n) to the blog's templates,
allowing for multi-language support and a more inclusive user experience.

The changes involve wrapping text in templates with the `{% trans %}` tag,
enabling the translation of static text into different languages based on the user's locale.

This enhancement ensures the platform is accessible to a wider audience
and improves the overall user experience by providing content in their preferred language.